+++ kos/src/libdebuginfo/debug_aranges.h
@@ -43,7 +43,7 @@ DECL_BEGIN
 INTDEF NONNULL((1, 2, 3)) unsigned int
 NOTHROW_NCX(CC libdi_debugaranges_locate)(byte_t *__restrict debug_aranges_start,
                                           byte_t *__restrict debug_aranges_end,
-                                          uint32_t *__restrict pdebug_info_cu_offset,
+                                          uintptr_t *__restrict pdebug_info_cu_offset,
                                           uintptr_t module_relative_pc);
